requirements.txt
Django==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
black==24.2.0
opencv-python==4.8.1.78
face_recognition==1.3.0